wasm dummy build
