+++ crates/grida-canvas/src/window/state.rs
@@ -2,14 +2,6 @@ use skia_safe::{
     Surface,
     gpu::{self, DirectContext, gl::FramebufferInfo},
 };
-use std::boxed::Box;
-
-#[cfg(target_arch = "wasm32")]
-unsafe extern "C" {
-    pub fn emscripten_GetProcAddress(
-        name: *const ::std::os::raw::c_char,
-    ) -> *const ::std::os::raw::c_void;
-}
 
 pub struct GpuState {
     pub context: DirectContext,
@@ -35,43 +27,50 @@ impl State {
     }
 }
 
-/// Load GL functions pointers from JavaScript so we can call OpenGL functions from Rust.
-///
-/// This only needs to be done once.
-#[cfg(target_arch = "wasm32")]
-fn init_gl() {
-    unsafe {
-        gl::load_with(|addr| {
-            let addr = std::ffi::CString::new(addr).unwrap();
-            emscripten_GetProcAddress(addr.into_raw() as *const _) as *const _
-        });
+impl State {
+    /// Create a new [`State`] for native targets using the provided context,
+    /// framebuffer information and surface. This mirrors the initialization
+    /// logic used on the wasm side.
+    pub fn from_parts(
+        context: DirectContext,
+        framebuffer_info: FramebufferInfo,
+        surface: Surface,
+    ) -> Self {
+        let gpu_state = GpuState {
+            context,
+            framebuffer_info,
+        };
+        Self::new(gpu_state, surface)
+    }
+
+    /// Recreate the underlying surface with the given dimensions.
+    pub fn resize(&mut self, width: i32, height: i32) {
+        let surface = create_surface(&mut self.gpu_state, width, height);
+        self.set_surface(surface);
+    }
+
+    /// Get a mutable pointer to the underlying Skia surface.
+    pub fn surface_mut_ptr(&mut self) -> *mut Surface {
+        &mut self.surface as *mut Surface
+    }
+
+    /// Borrow the underlying Skia surface mutably.
+    pub fn surface_mut(&mut self) -> &mut Surface {
+        &mut self.surface
+    }
+
+    /// Borrow the internal [`DirectContext`].
+    pub fn context(&mut self) -> &mut DirectContext {
+        &mut self.gpu_state.context
+    }
+
+    /// Retrieve framebuffer information.
+    pub fn framebuffer_info(&self) -> FramebufferInfo {
+        self.gpu_state.framebuffer_info
     }
 }
-
-/// Create the GPU state from the JavaScript WebGL context.
-///
-/// This needs to be done once per WebGL context.
-fn create_gpu_state() -> GpuState {
-    let interface = skia_safe::gpu::gl::Interface::new_native().unwrap();
-    let context = skia_safe::gpu::direct_contexts::make_gl(interface, None).unwrap();
-    let framebuffer_info = {
-        let mut fboid: gl::types::GLint = 0;
-        unsafe { gl::GetIntegerv(gl::FRAMEBUFFER_BINDING, &mut fboid) };
-
-        FramebufferInfo {
-            fboid: fboid.try_into().unwrap(),
-            format: skia_safe::gpu::gl::Format::RGBA8.into(),
-            protected: skia_safe::gpu::Protected::No,
-        }
-    };
-
-    GpuState {
-        context,
-        framebuffer_info,
-    }
-}
-
-fn create_surface(gpu_state: &mut GpuState, width: i32, height: i32) -> Surface {
+/// Create a GPU-backed surface used for rendering.
+pub(crate) fn create_surface(gpu_state: &mut GpuState, width: i32, height: i32) -> Surface {
     let backend_render_target =
         gpu::backend_render_targets::make_gl((width, height), 1, 8, gpu_state.framebuffer_info);
 
@@ -84,27 +83,4 @@ fn create_surface(gpu_state: &mut GpuState, width: i32, height: i32) -> Surface 
         None,
     )
     .unwrap()
-}
-
-/// Initialize the renderer.
-///
-/// This is called from JS after the WebGL context has been created.
-#[cfg(target_arch = "wasm32")]
-#[no_mangle]
-pub extern "C" fn init(width: i32, height: i32) -> Box<State> {
-    let mut gpu_state = create_gpu_state();
-    let surface = create_surface(&mut gpu_state, width, height);
-    let state = State::new(gpu_state, surface);
-    Box::new(state)
-}
-
-/// Resize the Skia surface
-///
-/// This is called from JS when the window is resized.
-#[cfg(target_arch = "wasm32")]
-#[no_mangle]
-pub unsafe extern "C" fn resize_surface(state: *mut State, width: i32, height: i32) {
-    let state = unsafe { state.as_mut() }.expect("got an invalid state pointer");
-    let surface = create_surface(&mut state.gpu_state, width, height);
-    state.set_surface(surface);
 }
