+++ crates/grida-math2/Cargo.toml
@@ -2,7 +2,7 @@
 [package]
 name = "math2"
 version = "0.0.2"
-edition = "2024"
+edition = "2021"
 description = "Geometry, layout and rasterization utilities powering the Grida canvas"
 license = "MIT"
 repository = "https://github.com/gridaco/grida"
